requirements.txt
discord.py
tensorflow==1.14
gpt-2-simple